Refactor groupby ops to avoid warnings
This has to do with a change in the groupby-apply behaviour of pandas.
In the older version, the groups passed to `apply` contained the
grouping columns. The new behaviour promotes removing the grouping
columns from the groups. For now it still allows both uses via the
`include_groups` argument, but it will become the default (or even
fixed) behaviour in the future.

Before the function used in `apply` (let's say `drop`) was returning a
dataframe that contained the grouping columns. `apply` would then
combine the resulting dataframes using the grouping columns as index.
Finally, we were resetting and dropping the _entire index_ because the
grouping columns were already in the dataframe.

Now, the grouping columns are not passed to `drop`, so the dataframe it
returns does not contain them. When `apply` combines these dataframes
using the grouping columns, we must preserve the index in order not to
lose this information. However, the dataframe produced in `drop`
contains another (meaningless) index. This is just because dataframes
must have an index. What we do in this case is to just drop the
meaningless index and restore the grouping columns which have been set
as the index.